Implemented read method
Created a read method to be called when the user selects option r in the prompt. This will give the user the option to print by title, or if they want to sort the tasks in a specific way then print out. Later on when the sort functions are added this section will also be updated