Implemented sorting by high priority
Added functionality for sorting tasks by highest priority to lowest. Also changed some of the prompt to change how the user is asked to print tasks, or be asked to sort the tasks. Also added test data in the main code so that I don't have to add tasks every time for testing