Fixed updating tasks not functioning properly
Updating tasks was not properly updating tasks because it was created before I began using the extra arrays such as priorities, due dates, and titles. Before when a task was updated, it would continue to use the previous task data in the priorities, due dates, and titles arrays creating confusion within the sort methods. I also updated the way the program prompts the user for input. With the old method update would throw an error because technically the method was still running, and once the data got changed it would throw a task not found error even though it existed.