Implemented auto import from file
Created an import from file function that will check for an existing SavedTasks.json file and if it exists will clear out the arrays the program uses and properly import all of the data for use. I have it set up now so that it automatically checks for a file to import. Also modified the export function a little to change the formatting in order to be used properly with the import function.